falcons/action_planning: refactor python test tooling, now we also have a way to quickly print summaries of all action .bin files
