+++ components/falcons/action_planning/test/BUILD
@@ -1,3 +1,15 @@
+py_library(
+    name = "action",
+    srcs = [
+        "action.py",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//components/falcons/action_planning/interface:interface_py",
+        "//libraries/logging:logdata",
+        "//libraries/plotting",
+    ],
+)
 
 py_binary(
     name = "plot",
@@ -7,8 +19,18 @@ py_binary(
     python_version = "PY3",
     visibility = ["//visibility:public"],
     deps = [
-        "//components/falcons/action_planning/interface:interface_py",
-        "//libraries/logging:logdata",
-        "//libraries/plotting",
+        ":action",
+    ],
+)
+
+py_binary(
+    name = "print",
+    srcs = [
+        "print.py",
+    ],
+    python_version = "PY3",
+    visibility = ["//visibility:public"],
+    deps = [
+        ":action",
     ],
 )
